pyproject.toml
[tool.poetry]
name = "carla-drl"
version = "1.0.1"
description = ""
authors = [
        "Dzhavid Sadreddinov <d.sadreddinov@innopolis.university>", 
        "Pavel Sharygin <p.sharygin@innopolis.university>",
        "Diana Vostrova <d.vostrova@innopolis.university>",
        "Alsu Khairullina <a.khairullina@innopolis.university>"
        ]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.7.12"
torch = {url = "https://download.pytorch.org/whl/cu116/torch-1.12.0%2Bcu116-cp37-cp37m-linux_x86_64.whl"}
torchvision = {url = "https://download.pytorch.org/whl/cu116/torchvision-0.13.0%2Bcu116-cp37-cp37m-linux_x86_64.whl"}
numpy = "1.18.4"
pillow = "9.5.0"
pygame = "2.6.1"
tensorboard = "2.11.2"
torch-tb-profiler = "0.4.3"
carla = {file = "carla/dist/carla-0.9.14-cp37-cp37m-manylinux_2_27_x86_64.whl"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-pylint = "^0.21.0"
pytest-mypy = "^0.10.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"